pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "neo4j-graphrag-kg"
version = "0.1.0"
description = "Neo4j-based Knowledge Graph for GraphRAG pipelines"
requires-python = ">=3.11"
dependencies = [
    "neo4j>=5.0",
    "typer>=0.9",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.4",
    "mypy>=1.0",
]

[project.scripts]
kg = "neo4j_graphrag_kg.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: marks tests that require a running Neo4j instance",
]
